Fixed some merge mistakes
